pw_toolchain: Fix alink tool for Windows
The recent commit removed the use of `rm` command before executing the
`arm-none-eabi-ar` tool due to the missing of `rm` executable on
Windows (661e85e).

This behaviour is not correct since `arm-none-eabi-ar` does not always
override the output file which may result in undesired content.

This commit fixes this issue by executing the `del` command. Due to
the fact of how `command` is executed on Windows, it is necessary to
append the `cmd`. The potential maximum command length limit issue is
addressed by using `rspfile` and `rspcontent`.

Change-Id: Ic2c28a73255aa39f43e1d31383d672f43feadfcc